modify example readme and wordtag use int64 for windows compatibility

+++ examples/information_extraction/wordtag/predictor.py
@@ -82,9 +82,9 @@ class WordtagPredictor(object):
         infer_ds = MapDataset(infer_data)
 
         batchify_fn = lambda samples, fn=Tuple(
-            Pad(axis=0, pad_val=self._tokenizer.pad_token_id),  # input_ids
-            Pad(axis=0, pad_val=self._tokenizer.pad_token_type_id),  # token_type_ids
-            Stack(),  # seq_len
+            Pad(axis=0, pad_val=self._tokenizer.pad_token_id,dtype='int64'),  # input_ids
+            Pad(axis=0, pad_val=self._tokenizer.pad_token_type_id,dtype='int64'),  # token_type_ids
+            Stack(dtype='int64'),  # seq_len
         ): fn(samples)
 
         infer_data_loader = paddle.io.DataLoader(
